nex-dev/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nex-dev"
version = "0.0.1"
description = "Utilities for handling python package development."
requires-python = ">=3.10"
dependencies = [
    "build>=1.2.1",
    "click>=8.1",
    "keyring>=25.1.0",
    "keyrings.google-artifactregistry-auth>=1.1.2",
    "google-cloud-artifact-registry>=1.11.3",
    "semver>=3.0.2",
    "tomlkit>=0.12.4",
    "twine>=5.0.0",
]

[project.entry-points.'nex-dev.subcommands']
package = "nex_dev.package:cli"
discover = "nex_dev.discover:cli"

[project.scripts]
nex-dev = "nex_dev:main"
